test(buildrequest): move resolver tests to fallback_resolve_test.go
Move the TestResolveFallbackChain_* cluster (24 tests, ~970 LOC) from
orchestrator_test.go to a new fallback_resolve_test.go that mirrors the
new fallback_resolve.go production seam. Verbatim move with no test
behaviour change; the tests already used package-internal symbols
(mockAdapter is in resolve_test.go, ClientRegistry literals built
in-test) so cross-file resolution Just Works in the same package.

orchestrator_test.go shrinks from 3148 to ~2180 LOC, retaining only
RunStreamOrchestration runtime tests sharing httptest-server fixtures.

Part of #199 item 4.
